requirements.txt
Django
djangorestframework
Pillow
django-heroku